+++ app/build.gradle
@@ -19,12 +19,11 @@ android {
 }
 
 dependencies {
-    implementation fileTree(include: ['*.jar'], dir: 'libs')
+    implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation 'com.android.support:appcompat-v7:28.0.0-alpha3'
     implementation 'com.android.support.constraint:constraint-layout:1.1.2'
-    implementation 'cz.msebera.android:httpclient:4.4.1.1'
+    implementation group: 'cz.msebera.android' , name: 'httpclient', version: '4.4.1.1'
     testImplementation 'junit:junit:4.12'
     androidTestImplementation 'com.android.support.test:runner:1.0.2'
     androidTestImplementation 'com.android.support.test.espresso:espresso-core:3.0.2'
-    implementation files('libs/YouTubeAndroidPlayerApi.jar')
 }
