+++ app/src/main/AndroidManifest.xml
@@ -21,8 +21,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".GameActivity"
-            android:hardwareAccelerated="true"/>
         <provider
             android:name="android.support.v4.content.FileProvider"
             android:authorities="com.example.provider"
