+++ app/src/main/res/values/strings.xml
@@ -3,9 +3,8 @@
     <string name="happiness">Happiness</string>
     <string name="sadness">Sadness</string>
     <string name="anger">Anger</string>
-    <string name="Start">Start</string>
+    <string name="Camera">Take Photo</string>
     <string name="Neutral">Neutral</string>
     <string name="Age">Age</string>
     <string name="Gender">Gender</string>
-    <string name="GameTitle">You laugh  You Lose</string>
 </resources>
